chore(client): move highway into internal/highway
进一步将目录结构与MiraiGo对齐
